backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
python-dotenv==0.19.0
djangorestframework-simplejwt==4.7.2
django-filter~=21.1
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
djoser==2.1.0
django-colorfield=0.7.2